Started setting up location in StartActivity class and removed location check method in MapsActivity. Work in progress
