+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".MapsActivity"
+            android:name=".StartActivity"
             android:label="@string/title_activity_maps">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -45,7 +45,8 @@
         <activity
             android:name=".PubActivity"
             android:label="@string/title_activity_pub" />
-        <activity android:name=".PubCompareActivity"></activity>
+        <activity android:name=".PubCompareActivity" />
+        <activity android:name=".MapsActivity"></activity>
     </application>
 
 </manifest>
