+++ app/src/main/res/layout/activity_start.xml
@@ -6,4 +6,14 @@
     android:layout_height="match_parent"
     tools:context=".StartActivity">
 
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Start"
+        android:id="@+id/startButton"
+        />
+
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
